requirements.txt
annotated-types==0.6.0
anyio==3.7.1
astroid==2.15.8
attrs==23.2.0
autopep8==2.0.1
cachetools==5.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.2.1
cryptography==42.0.5
dill==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.103.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
orjson==3.10.1
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.5
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-mock==3.10.0
pytest-reportlog==0.4.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
requests==2.28.2
rsa==4.9
ruff==0.3.7
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.20
sqlalchemy-stubs==0.4
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
